Update README.md
Loan classification is one of the oldest problems in the world of Machine Learning. Here Learning how to predict if a person will be able to pay the loan with logistic regression algorithm using sklearn library for machine learning.

This project concentrates more on data cleansing and working data transformation techniques like quantiles, fixing distributions, etc., along with detecting outliers.